requirements.txt
gallery-dl
aria2
google-generativeai
openai
googletrans-py